// src/api/generated/model/draftCreateRequestEntityType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Fitness App API
 * Comprehensive API for fitness application with support for workout plans, exercises, and progress tracking. API provides OAuth2 authentication via Keycloak or JWT tokens.
 */

export type DraftCreateRequestEntityType = typeof DraftCreateRequestEntityType[keyof typeof DraftCreateRequestEntityType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DraftCreateRequestEntityType = {
  PLAN: 'PLAN',
  WORKOUT: 'WORKOUT',
  WORKOUT_TEMPLATE: 'WORKOUT_TEMPLATE',
} as const;