// src/api/generated/model/workoutExerciseInputRequestWorkoutExerciseMetric.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Fitness App API
 * Comprehensive API for fitness application with support for workout plans, exercises, and progress tracking. API provides OAuth2 authentication via Keycloak or JWT tokens.
 */

export type WorkoutExerciseInputRequestWorkoutExerciseMetric = typeof WorkoutExerciseInputRequestWorkoutExerciseMetric[keyof typeof WorkoutExerciseInputRequestWorkoutExerciseMetric];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WorkoutExerciseInputRequestWorkoutExerciseMetric = {
  REPS: 'REPS',
  TIME: 'TIME',
  REPS_AND_WEIGHT: 'REPS_AND_WEIGHT',
  TIME_AND_WEIGHT: 'TIME_AND_WEIGHT',
  DISTANCE: 'DISTANCE',
  DISTANCE_AND_TIME: 'DISTANCE_AND_TIME',
} as const;